退出時のバグを修正
